requirements.txt
google-generativeai
requests
python-dotenv
streamlit